Update README.md
Demon Slayer: Mugen Train (2020)

Plot: Demon Slayer: Mugen Train is the first movie in the Demon Slayer series and serves as a direct continuation of the anime’s first season. The movie adapts the Mugen Train arc from the manga.

 Demon Slayer: The Entertainment District Arc (2021)

While this is technically a follow-up to the Mugen Train film and not a standalone movie, this arc was adapted into the anime series with some movie-like production value. The Entertainment District Arc introduces a new set of challenges as the Demon Slayer Corps goes up against powerful demons in a red-light district.